feat: jsdoc support
