+++ src/@rocket-scripts/web/index.ts
@@ -1,2 +1,3 @@
+export * from './params';
 export * from './start';
 export * from './build';
